WARNING !!!VERY MESSY!!! TLDR: Added a new button class to the main file for the upgrade buttons.
new button class is in the main file in the outermost scope. Added this for text buttons (like the upgrade buttons). My other change was that a tower remains selected if the bottom menu or pause button is clicked. I might see if I can put this class into another file so it isn't so messy. Currently the buttons have no functionality but clicks are registered